fix: linter errors

+++ src/services/user.rs
@@ -134,24 +134,21 @@ impl RegistrationService {
                     drop(self.user_repository.grant_admin_role(&user_id).await);
                 }
 
-                match &registration.email {
-                    Some(email) => {
-                        if email.verification_required {
-                            // Email verification is enabled
-                            if let Some(email) = opt_email {
-                                let mail_res = self
-                                    .mailer
-                                    .send_verification_mail(&email, &registration_form.username, user_id, api_base_url)
-                                    .await;
+                if let Some(email) = &registration.email {
+                    if email.verification_required {
+                        // Email verification is enabled
+                        if let Some(email) = opt_email {
+                            let mail_res = self
+                                .mailer
+                                .send_verification_mail(&email, &registration_form.username, user_id, api_base_url)
+                                .await;
 
-                                if mail_res.is_err() {
-                                    drop(self.user_repository.delete(&user_id).await);
-                                    return Err(ServiceError::FailedToSendVerificationEmail);
-                                }
+                            if mail_res.is_err() {
+                                drop(self.user_repository.delete(&user_id).await);
+                                return Err(ServiceError::FailedToSendVerificationEmail);
                             }
                         }
                     }
-                    None => (),
                 }
 
                 Ok(user_id)
